query_agents/requirements/stable.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.23
chromedriver-autoinstaller==0.6.4
click==8.1.7
coloredlogs==15.0.1
courlan==1.0.0
cssselect==1.2.0
dataclasses-json==0.6.4
dateparser==1.2.0
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
html2text==2020.1.16
htmldate==1.7.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
jieba3k==0.35.1
joblib==1.3.2
jusText==3.0.0
kubernetes==29.0.0
langcodes==3.3.0
llama-index==0.10.12
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.5
llama-index-core==0.10.12
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.6
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.5
llama-index-readers-github==0.1.7
llama-index-readers-llama-parse==0.1.3
llama-index-readers-web==0.1.6
llama-index-vector-stores-chroma==0.1.4
llama-parse==0.3.4
llamaindex-py-client==0.1.13
lxml==5.1.0
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.14
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
python-dotenv
pandas==2.2.0
pillow==10.2.0
playwright==1.41.2
posthog==3.4.2
protobuf==4.25.3
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.0.1
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pypdf==4.0.2
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.18.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.1
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
trafilatura==1.7.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
